emacs: better support for draw.io files
Instead of switching between browser-based draw.io and org-mode notes,
now it is possible to create a diagram in a few key strokes and
refer back to it later from its link.

I don't know how easy it is to manage attachments when moving nodes
between files or moving attachment links, but it's not really a big deal
right now.