emacs: In org, open new attachments in new frame
Most of the time I'm going to want to copy some snippet from the
attachment or refer back to it when writing a note. Having it open in a
separate frame does that. It also makes it easier to just `:wq` to save
and exit from it compared to `:w` and switching back to my original
buffer.